If sys calls fail in CGI, we don't quit running the server

+++ src/CgiHandler.cpp
@@ -1,10 +1,15 @@
 #include "../inc/CgiHandler.hpp"
+#include "../inc/Signal.hpp"
 #include <stdexcept>
 
 t_cgi_state	CgiHandler::checkCgi()
 {
 	int	status;
 
+	if (_state == CGI_FAILED)
+	{
+		return (CGI_FAILED);
+	}
 	if (std::time(nullptr) - _start_time >= _CGI_TIMEOUT_S){
 		std::cout << "KILL THE CHILDREN\n";
 		kill(_pid, SIGKILL);
@@ -109,22 +114,12 @@ void	CgiHandler::launchCgi(Request &request)
 
 	int	output_fd = open(output_filename.c_str(), O_RDWR | O_CREAT | O_TRUNC, 0644); // let's verify permissions later
 	if (output_fd < 0){
-		// close(input_fd);
-		// throw
+		_state = CGI_FAILED;
+		throw std::runtime_error("Failed to open file in launchCgi");
 	}
-	
 	getCgiEnv(request, request.getConfig(), envp, envVars);
-	// we have saved the post body into infile before
-	// close(input_fd); // move file position indicator to the beginning of the file stream by closing
 
-	std::string	executable;
-	if (endsWith(request.getPath(), ".py")) {
-		executable = request.getLocation().cgi_path_py;
-	} else if (endsWith(request.getPath(), ".bla")) {
-		executable = request.getLocation().cgi_path_bla;
-	} else {
-		throw std::runtime_error("launchCgi(): request path ending unsupported");
-	}
+	std::string	executable = request.getLocation().cgi_path_py;
 	std::string	script = request.getLocation().root + request.getPath();
 
 	// std::cout << "\n\nCGI_PATH: " << executable << "\n\nrequest.path: " << request.getPath() << "\n\nSCRIPT_PATH: " << script << std::endl;
@@ -141,7 +136,7 @@ void	CgiHandler::launchCgi(Request &request)
 	if (pid < 0) {
 		// close(input_fd);
 		close(output_fd);
-		throw std::runtime_error("fork() failed"); // are we closing the server?
+		throw std::runtime_error("fork() failed");
 	}
 	else if (pid == 0) {
 		int input_fd = open(request.getPostBodyFilename().c_str(), O_RDONLY, 0644);
@@ -153,7 +148,8 @@ void	CgiHandler::launchCgi(Request &request)
 		dup2(output_fd, STDOUT_FILENO);
 		close(output_fd);
 		execve(argv[0], argv, envp.data());
-		// throw something ?
+		_state = CGI_FAILED;
+		throw std::runtime_error("CGI: execve failed");
 	}
 	else {
 		_start_time = std::time(nullptr);
